frontend/requirements.txt
requests
streamlit==0.83.0
Pillow
plotly
pandas 
